store cart_item

+++ app/Http/Controllers/FrontController.php
@@ -3,10 +3,10 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Gallery;
-use App\Cart;
 use Illuminate\Support\Facades\Auth;
 use DB;
+use App\Gallery;
+use App\Cart;
 use App\Address;
 use App\Order;
 use App\Order_detail;
@@ -20,6 +20,7 @@ class FrontController extends Controller
          $galleries = Gallery::orderBy('created_at', 'desc')->where('sold_out',1)->paginate(6);
         return view('frontend.gallery', compact("galleries"));
     }
+
     public function detailGallery($id)
     {
     	$galleries = Gallery::where('id', $id)
@@ -27,16 +28,9 @@ class FrontController extends Controller
            
 		return view('frontend.detail', compact('galleries'));
     }
-    public function addtocart(Request $request)
-    {
-        $cart = new Cart();
-        $cart->user_id = Auth::user()->id;
-        $cart->gallery_id = $request->gallery_id;
-        $cart->save();
     
-         return redirect('/cart');
-    }
-    public function address(Request $request){
+    public function address(Request $request)
+    {
         $validatedData = $request->validate([
             'phone' => 'required',
             'address' => 'required',
@@ -45,7 +39,7 @@ class FrontController extends Controller
 
         Auth::user()->address()->create($validatedData);
 
-      return redirect('/orderConfirm');
+        return redirect('/orderConfirm');
     }
     
     public function orderConfirm(Request $request)
@@ -54,27 +48,45 @@ class FrontController extends Controller
             ->latest()
             ->first();
 
-        $carts = DB::table('galleries')
-                ->leftJoin('carts', 'galleries.id', "=", 'carts.gallery_id')
-                ->where('user_id', Auth::user()->id)
+       /*  $carts = Cart_item::with(['galleries','cart'])->where()->get();*/
+
+        $cart=Cart::where('user_id',Auth::user()->id)->first();
+      
+        $carts = DB::table('cart_items')
+                ->leftJoin('galleries','cart_items.gallery_id',"=",'galleries.id')
+                ->leftJoin('carts','cart_items.cart_id',"=",'carts.id')
+                ->where('carts.user_id',$cart->user_id)
                 ->get();
+
+        $quantities = $carts->count();
+        $total = $carts->sum('price');
+
+          session([ 
+            'total' => $total,
+            'quantities' => $quantities,
+        ]);
 
         return view('frontend.order_confirm', compact('shipping_address','carts'));
     }
+
     public function storeOrder(Request $request)
     {
-         $order=DB::table('orders')->where('user_id', Auth::user()->id)->first(); 
+        $cart=Cart::where('user_id',Auth::user()->id)->first();
 
-         $cart =  DB::table('galleries')
-            ->leftJoin('carts', 'galleries.id', "=", 'carts.gallery_id')
-            ->where('user_id', Auth::user()->id)
-            ->get();
+        $carts = DB::table('cart_items')
+                ->leftJoin('galleries','cart_items.gallery_id',"=",'galleries.id')
+                ->leftJoin('carts','cart_items.cart_id',"=",'carts.id')
+                ->where('carts.user_id',$cart->user_id)
+                ->get();
+
+        $order = Order::with('user')->where('user_id', Auth::user()->id)->first(); 
+
         $order = Order::create([
             'user_id' => Auth::user()->id,
-            'payment_method'=>1,
+           'payment_method'=>1,
         ]);
 
-        $cart->each(function ($item) use ($order) {
+        $carts->each(function ($item) use ($order) {
             Order_detail::create([
                 'gallery_id' => $item->gallery_id,
                 'order_id' => $order->id,
@@ -82,11 +94,29 @@ class FrontController extends Controller
             
         });
 
-        $cart_delete = Cart::where('user_id', Auth::user()->id);   
-            $cart_delete->delete();
+        $cart_delete = Cart::where('user_id',Auth::user()->id)->delete();
+        $cart_item=DB::table('cart_items')->where('cart_id', $cart->id)->delete();
 
+      
         $request->session()->forget(['total', 'quantities']);
         return view('frontend.thankyou');
+    }
+    
+    public function storeCart(Request $request)
+    {
+
+        $cart = Cart::with('user')->where('user_id', Auth::user()->id)->first(); 
+
+        $cart = Cart::create([
+            'user_id' => Auth::user()->id, 
+        ]);
+    
+        Cart_item::create([
+                'gallery_id' => $request->gallery_id,
+                'cart_id' => $cart->id,
+            ]);
+
+       return redirect('/cart');
     }
     
 }
